Room Generation fixes in progress
All 4 directions are checked with raycasts.

Before picking a direction to Move(), raycasts check which 4 directions are open.
Those open directions are added to a list, a random value is picked from that list to Move().
If no directions are open, a random generated room is picked and the directions are picked again.

Fixed issue with raycast hitting itself. Disabled in Project Settings -> Physics2D -> Queries Start In Colliders